requirements.txt
aiogram==3.3.0

psycopg2-binary

psycopg2-binary
